requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.10
cycler==0.11.0
fonttools==4.29.0
idna==3.3
kiwisolver==1.3.2
matplotlib==3.5.1
nba-api==1.1.11
numpy==1.22.1
packaging==21.3
pandas==1.4.0
Pillow==9.0.0
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
requests==2.27.1
six==1.16.0
torch==1.10.1
torchvision==0.11.2
typing_extensions==4.0.1
urllib3==1.26.8
